### 2017-09-14 aa3a458
Added a couple of catch statement that make sure the energy range defined is correct.
