+++ tests/ImagingReso/resonance_test.py
@@ -1,24 +1,23 @@
 import unittest
 import numpy as np
-import os
-import pprint
 
 from ImagingReso.resonance import Resonance
+
 
 class TestInitialization(unittest.TestCase):
 
     def test_E_min(self):
-        '''assert E can not be set below a given threshold'''
+        """assert E can not be set below a given threshold"""
         energy_min = 1e-6
         self.assertRaises(ValueError, Resonance, energy_min=energy_min)
 
     def test_E_max(self):
-        '''assert E can not be set above a given threshold'''
+        """assert E can not be set above a given threshold"""
         energy_max = 10000
         self.assertRaises(ValueError, Resonance, energy_max=energy_max)
 
     def test_str(self):
-        '''assert print(object) works'''
+        """assert print(object) works"""
         _stack = {'CoAg': {'elements': ['Co','Ag'],
                                'stoichiometric_ratio': [1, 2],
                                    'thickness': {'value': 0.025,
@@ -38,7 +37,7 @@ class TestInitialization(unittest.TestCase):
         self.assertIsInstance(o_reso.__str__(), str)
 
     def test_repr(self):
-        '''assert object works'''
+        """assert object works"""
         _stack = {'CoAg': {'elements': ['Co','Ag'],
                                'stoichiometric_ratio': [1, 2],
                                    'thickness': {'value': 0.025,
@@ -58,7 +57,7 @@ class TestInitialization(unittest.TestCase):
         self.assertIsInstance(o_reso.__repr__(), str)
         
     def test_initialization_of_object(self):
-        '''assert object is correctly initialized with energy min, max and step'''
+        """assert object is correctly initialized with energy min, max and step"""
         _stack = {'CoAg': {'elements': ['Co','Ag'],
                                'stoichiometric_ratio': [1, 2],
                                'thickness': {'value': 0.025,
@@ -79,8 +78,46 @@ class TestInitialization(unittest.TestCase):
         self.assertEqual(o_reso.energy_min, energy_min)
         self.assertEqual(o_reso.energy_step, energy_step)
 
+    def test_initialization_with_same_E_min_and_max_raises_error(self):
+        """assert ValueError is raised if E_min == E_max"""
+        _stack = {'CoAg': {'elements': ['Co', 'Ag'],
+                           'stoichiometric_ratio': [1, 2],
+                           'thickness': {'value': 0.025,
+                                         'units': 'mm'},
+                           },
+                  'Ag': {'elements': ['Ag'],
+                         'stoichiometric_ratio': [1],
+                         'thickness': {'value': 0.03,
+                                       'units': 'mm'},
+                         },
+                  }
+        energy_min = 150
+        energy_max = 150
+        energy_step = 1
+        self.assertRaises(ValueError, Resonance, stack=_stack, energy_max=energy_max, energy_min=energy_min,
+                           energy_step=energy_step)
+
+    def test_initialization_E_step_bigger_than_E_range_raises_error(self):
+        """assert ValueError is raised if E_min == E_max"""
+        _stack = {'CoAg': {'elements': ['Co', 'Ag'],
+                           'stoichiometric_ratio': [1, 2],
+                           'thickness': {'value': 0.025,
+                                         'units': 'mm'},
+                           },
+                  'Ag': {'elements': ['Ag'],
+                         'stoichiometric_ratio': [1],
+                         'thickness': {'value': 0.03,
+                                       'units': 'mm'},
+                         },
+                  }
+        energy_min = 1
+        energy_max = 10
+        energy_step = 11
[…]
