+++ tests/ImagingReso/resonance_test.py
@@ -18,17 +18,17 @@ class TestInitialization(unittest.TestCase):
 
     def test_str(self):
         """assert print(object) works"""
-        _stack = {'CoAg': {'elements': ['Co','Ag'],
-                               'stoichiometric_ratio': [1, 2],
-                                   'thickness': {'value': 0.025,
-                                                 'units': 'mm'},
-                                   },
-                      'Ag': {'elements': ['Ag'],
-                                 'stoichiometric_ratio': [1],
-                                 'thickness': {'value': 0.03,
-                                               'units': 'mm'},
-                                 },
-                          }
+        _stack = {'CoAg': {'elements': ['Co',' Ag'],
+                           'stoichiometric_ratio': [1, 2],
+                           'thickness': {'value': 0.025,
+                                         'units': 'mm'},
+                           },
+                  'Ag': {'elements': ['Ag'],
+                         'stoichiometric_ratio': [1],
+                         'thickness': {'value': 0.03,
+                                       'units': 'mm'},
+                         },
+                  }
         energy_min = 10
         energy_max = 150
         energy_step = 1
@@ -38,17 +38,17 @@ class TestInitialization(unittest.TestCase):
 
     def test_repr(self):
         """assert object works"""
-        _stack = {'CoAg': {'elements': ['Co','Ag'],
-                               'stoichiometric_ratio': [1, 2],
-                                   'thickness': {'value': 0.025,
-                                                 'units': 'mm'},
-                                   },
-                      'Ag': {'elements': ['Ag'],
-                                 'stoichiometric_ratio': [1],
-                                 'thickness': {'value': 0.03,
-                                               'units': 'mm'},
-                                 },
-                          }
+        _stack = {'CoAg': {'elements': ['Co', 'Ag'],
+                           'stoichiometric_ratio': [1, 2],
+                           'thickness': {'value': 0.025,
+                                         'units': 'mm'},
+                           },
+                  'Ag': {'elements': ['Ag'],
+                         'stoichiometric_ratio': [1],
+                         'thickness': {'value': 0.03,
+                                       'units': 'mm'},
+                         },
+                  }
         energy_min = 10
         energy_max = 150
         energy_step = 1
@@ -146,7 +146,7 @@ class TestInitialization(unittest.TestCase):
 
     def test_correct_initialization_of_stack(self):
         """assert correct defined stack is correctly saved"""
-        _stack = {'CoAg': {'elements': ['Co','Ag'],
+        _stack = {'CoAg': {'elements': ['Co', 'Ag'],
                            'stoichiometric_ratio': [1, 2],
                            'thickness': {'value': 0.025,
                                          'units': 'mm'},
@@ -178,14 +178,14 @@ class TestInitialization(unittest.TestCase):
         
         returned_stack = o_reso.stack
         
-        expected_stack = {'CoAg': {'elements': ['Co','Ag'],
-                                   'stoichiometric_ratio': [1,1],
+        expected_stack = {'CoAg': {'elements': ['Co', 'Ag'],
+                                   'stoichiometric_ratio': [1, 1],
                                    'thickness': {'value': 0.025,
                                                  'units': 'mm'},
                                    'density': {'value': 9.7,
                                                'units': 'g/cm3'},
-                                   'Co': {'isotopes': {'file_names': ['Co-58.csv','Co-59.csv'],
-                                                       'list': ['58-Co','59-Co'],
+                                   'Co': {'isotopes': {'file_names': ['Co-58.csv', 'Co-59.csv'],
+                                                       'list': ['58-Co', '59-Co'],
[…]
