### 2017-09-14 f62a9c2
cleaned up commented code.
