+++ ImagingReso/_utilities.py
@@ -24,12 +24,13 @@ def is_element_in_database(element='', database='ENDF_VIII'):
     '''
     if element == '':
         return False
-    
+
     list_entry_from_database = get_list_element_from_database(database=database)
     if element.lower() in list_entry_from_database:
         return True
-    return False    
-    
+    return False
+
+
 def get_list_element_from_database(database=''):
     '''return a string array of all the element from the database
     
@@ -47,12 +48,13 @@ def get_list_element_from_database(database=''):
 
     if not os.path.exists(_database_folder):
         raise ValueError("Database {} does not exist!".format(database))
-    
+
     _list_files = glob.glob(_database_folder + '/*.csv')
     _list_short_files = [os.path.basename(_file) for _file in _list_files]
     _list_element = set([_name.split('-')[0].lower() for _name in _list_short_files])
     return _list_element
-    
+
+
 def checking_stack(stack={}, database='ENDF_VIII'):
     '''This method makes sure that all the elements from the various stacks are 
     in the database and that the thickness has the correct format (float)
@@ -81,12 +83,13 @@ def checking_stack(stack={}, database='ENDF_VIII'):
         _thickness = stack[_keys]['thickness']['value']
         if not isinstance(_thickness, numbers.Number):
             raise ValueError("Thickness {} should be a number!".format(_thickness))
-    
+
         _stoichiometric_ratio = stack[_keys]['stoichiometric_ratio']
         if len(_stoichiometric_ratio) != len(_elements):
             raise ValueError("stoichiometric Ratio and Elements should have the same size!")
-    
-    return True    
+
+    return True
+
 
 def formula_to_dictionary(formula='', thickness=np.NaN, density=np.NaN, database='ENDF_VIII'):
     '''create dictionary based on formula given
@@ -143,6 +146,7 @@ def formula_to_dictionary(formula='', thickness=np.NaN, density=np.NaN, database
                       },
             }
 
+
 def get_isotope_dicts(element='', database='ENDF_VIII'):
     '''return a dictionary with list of isotopes found in database and name of database files
     
@@ -163,19 +167,19 @@ def get_isotope_dicts(element='', database='ENDF_VIII'):
     _database_folder = os.path.join(_file_path, 'reference_data', database)
     _element_search_path = os.path.join(_database_folder, element + '-*.csv')
     list_files = glob.glob(_element_search_path)
-    isotope_dict = {'isotopes': {'list': [], 
+    isotope_dict = {'isotopes': {'list': [],
                                  'file_names': [],
                                  'density': {'value': np.NaN,
                                              'units': 'g/cm3'},
                                  'mass': {'value': [],
                                           'units': 'g/mol',
-                                 },
-                                 'isotopic_ratio': [],},
+                                          },
+                                 'isotopic_ratio': [], },
                     'density': {'value': np.NaN,
                                 'units': 'g/cm3'},
                     'molar_mass': {'value': np.NaN,
                                    'units': 'g/mol'},
-                    } 
+                    }
 
     isotope_dict_mirror = {}
[…]
