+++ ImagingReso/_utilities.py
@@ -10,7 +10,7 @@ from scipy.constants import Avogadro
 
 
 def is_element_in_database(element='', database='ENDF_VIII'):
-    '''will try to find the element in the folder (database) specified
+    """will try to find the element in the folder (database) specified
     
     Parameters:
     ==========
@@ -21,7 +21,7 @@ def is_element_in_database(element='', database='ENDF_VIII'):
     =======
     bool: True if element was found in the database
           False if element could not be found
-    '''
+    """
     if element == '':
         return False
 
@@ -32,7 +32,7 @@ def is_element_in_database(element='', database='ENDF_VIII'):
 
 
 def get_list_element_from_database(database=''):
-    '''return a string array of all the element from the database
+    """return a string array of all the element from the database
     
     Parameters:
     ==========
@@ -42,7 +42,7 @@ def get_list_element_from_database(database=''):
     ======
     ValueError if database can not be found
     
-    '''
+    """
     _file_path = os.path.abspath(os.path.dirname(__file__))
     _database_folder = os.path.join(_file_path, 'reference_data', database)
 
@@ -56,7 +56,7 @@ def get_list_element_from_database(database=''):
 
 
 def checking_stack(stack={}, database='ENDF_VIII'):
-    '''This method makes sure that all the elements from the various stacks are 
+    """This method makes sure that all the elements from the various stacks are 
     in the database and that the thickness has the correct format (float)
     
     Parameters:
@@ -73,7 +73,7 @@ def checking_stack(stack={}, database='ENDF_VIII'):
     Return:
     ======
     True: for testing purpose only
-    '''
+    """
     for _keys in stack:
         _elements = stack[_keys]['elements']
         for _element in _elements:
@@ -92,7 +92,7 @@ def checking_stack(stack={}, database='ENDF_VIII'):
 
 
 def formula_to_dictionary(formula='', thickness=np.NaN, density=np.NaN, database='ENDF_VIII'):
-    '''create dictionary based on formula given
+    """create dictionary based on formula given
     
     Parameters:
     ===========
@@ -119,7 +119,7 @@ def formula_to_dictionary(formula='', thickness=np.NaN, density=np.NaN, database
                                  'units': 'g/cm3',
                                  },
                     }
-    '''
+    """
     _formula_parsed = re.findall(r'([A-Z][a-z]*)(\d*)', formula)
 
     _dictionary = {}
@@ -148,7 +148,7 @@ def formula_to_dictionary(formula='', thickness=np.NaN, density=np.NaN, database
 
 
 def get_isotope_dicts(element='', database='ENDF_VIII'):
-    '''return a dictionary with list of isotopes found in database and name of database files
+    """return a dictionary with list of isotopes found in database and name of database files
     
     Parameters:
     ===========
@@ -162,7 +162,7 @@ def get_isotope_dicts(element='', database='ENDF_VIII'):
       ex: {'Ag': {'isotopes': ['107-Ag','109-Ag'],
                   'file_names': ['Ag-107.csv','Ag-109.csv']}}
     
-    '''
+    """
     _file_path = os.path.abspath(os.path.dirname(__file__))
     _database_folder = os.path.join(_file_path, 'reference_data', database)
     _element_search_path = os.path.join(_database_folder, element + '-*.csv')
@@ -218,7 +218,7 @@ def get_isotope_dicts(element='', database='ENDF_VIII'):
 
 
 def get_abundance(element):
[…]
