+++ ImagingReso/resonance.py
@@ -8,7 +8,6 @@ from ImagingReso import _utilities
 
 
 class Resonance(object):
-    
     database = 'ENDF_VIII'
 
     E_MIN = 1e-5
@@ -18,9 +17,9 @@ class Resonance(object):
     stack_sigma = {}  # all the energy and sigma of the isotopes and compounds
     stack_signal = {}  # transmission and attenuation signal for every isotope and compound
     total_signal = {}  # transmission and attenuation of the entire sample
-    
+
     density_lock = {}  # dictionary that will defined the densities locked
-    
+
     energy_max = np.NaN
     energy_min = np.NaN
     energy_step = np.NaN
@@ -67,13 +66,13 @@ class Resonance(object):
             _utilities.checking_stack(stack=stack)
             new_stack = self.__update_stack_with_isotopes_infos(stack=stack)
             self.stack = new_stack
-            
+
             # if layer density has been defined, lock it
             self.__lock_density_if_defined(stack=self.stack)
-            
+
             # calculate stack_sigma, layer density, atoms_per_cm3 ...
             self.__math_on_stack()
-              
+
     def __str__(self):
         '''what to display if user does
         
@@ -81,16 +80,16 @@ class Resonance(object):
         >>> print(o_reso)
         '''
         return json.dumps(self.stack, indent=4)
-    
+
     def __repr__(self):
         '''what to display if user does
         
         >>> o_reso = Resolution()
         >>> o_reso
         '''
-        return json.dumps(self.stack, indent=4)    
-                
-    def add_layer(self, formula='', thickness=np.NaN, density=np.NaN): 
+        return json.dumps(self.stack, indent=4)
+
+    def add_layer(self, formula='', thickness=np.NaN, density=np.NaN):
         """provide another way to define the layers (stack)
         
         Parameters:
@@ -103,17 +102,17 @@ class Resonance(object):
         """
         if formula == '':
             return
-        
-        _new_stack = _utilities.formula_to_dictionary(formula=formula, 
-                                                      thickness=thickness, 
+
+        _new_stack = _utilities.formula_to_dictionary(formula=formula,
+                                                      thickness=thickness,
                                                       density=density,
-                                                      database=self.database)      
+                                                      database=self.database)
         # check if density has been defined
         self.__lock_density_if_defined(stack=_new_stack)
-        
+
         new_stack = self.__update_stack_with_isotopes_infos(stack=_new_stack)
         self.stack = {**self.stack, **new_stack}
-        
+
         # calculate stack_sigma, layer density, atoms_per_cm3 ...
         self.__math_on_stack()
 
@@ -131,7 +130,7 @@ class Resonance(object):
         """
         _stack = self.stack
         compound = str(compound)
-        
+
         if compound == '':
             _list_compounds = _stack.keys()
             list_all_dict = {}
@@ -144,32 +143,32 @@ class Resonance(object):
                         compound=_compound,
                         element=_element)
             return list_all_dict
-        
+
         # checking compound is valid
         list_compounds = _stack.keys()
         if not compound in list_compounds:
             list_compounds_joined = ', '.join(list_compounds)
             raise ValueError("Compound '{}' could not be find in {}".format(compound, list_compounds_joined))
-        
+
         # checking element is valid
-        if element == '': 
+        if element == '':
             # we assume that the element and compounds names matched
             element = compound
         list_element = _stack[compound].keys()
         if not element in list_element:
             list_element_joined = ', '.join(list_element)
             raise ValueError("Element '{}' should be any of those elements: {}".format(element, list_element_joined))
-        
+
         list_istopes = _stack[compound][element]['isotopes']['list']
[…]

[…]
