### 2017-09-19 6ad3a2d
Started to make PEP 8 happy.
