### 2017-09-20 #407 @whedon
[REVIEW]: ImagingReso: A Tool for Neutron Resonance Imaging
**Submitting author:** @zhangy6x (<a href="http://orcid.org/0000-0002-0083-1408">Yuxuan Zhang</a>)
**Repository:** <a href="https://github.com/ornlneutronimaging/ImagingReso" target ="_blank">https://github.com/ornlneutronimaging/ImagingReso</a>
**Version:** v1.5.8
