### 2017-09-20 #407 @nicoguaro
@zhangy6x, I suggest that you make explicit the dependence on Python 3. I will need to repeat the process since I created an environment for Python 2.7 instead.
