### 2017-09-21 15afac9
addressed the support only for Python 3.x
