### 2017-09-21 #407 @zhangy6x
@nicoguaro: Good point, I just addressed the Python version supported in the README.rst.
