+++ tests/ImagingReso/utilities_test.py
@@ -9,33 +9,32 @@ from ImagingReso._utilities import *
 
 
 class TestUtilities_1(unittest.TestCase):
-    
     def test_is_element_in_database(self):
-        '''assert is_element_in_database works correctly for good and bad input elements'''
+        """assert is_element_in_database works correctly for good and bad input elements"""
 
         # empty element
         _element = ''
         _answer = is_element_in_database(element=_element)
         self.assertFalse(_answer)
-    
+
         # element not in database
         _element = 'Ne'
         _answer = is_element_in_database(element=_element)
         self.assertFalse(_answer)
-        
+
         # element in database
         _element = 'Co'
         _answer = is_element_in_database(element=_element)
         self.assertTrue(_answer)
-        
+
     def test_get_list_of_element_raise_error_if_wrong_database(self):
-        '''assert ValueError if wrong database passed to get_list_element_from_database'''
+        """assert ValueError if wrong database passed to get_list_element_from_database"""
         _database_1 = 'do_not_exist'
         self.assertRaises(ValueError, get_list_element_from_database, database=_database_1)
-        
+
     def test_checking_stack(self):
-        '''assert checking_stack_works in all good cases (1 or more stacks)'''
-        
+        """assert checking_stack_works in all good cases (1 or more stacks)"""
+
         # normal simple stack
         stack_1 = {'Ag': {'elements': ['Ag'],
                           'stoichiometric_ratio': [1],
@@ -45,128 +44,128 @@ class TestUtilities_1(unittest.TestCase):
                    }
         _result_1 = checking_stack(stack=stack_1)
         self.assertTrue(_result_1)
-        
+
         # complex set of stacks
         stack_2 = {'Co': {'elements': ['Co'],
                           'stoichiometric_ratio': [1],
                           'thickness': {'value': 0.03,
                                         'units': 'mm'},
                           },
-                   'GdEu': {'elements': ['Gd','Eu'],
-                            'stoichiometric_ratio': [1,1],
+                   'GdEu': {'elements': ['Gd', 'Eu'],
+                            'stoichiometric_ratio': [1, 1],
                             'thickness': {'value': 0.025,
                                           'units': 'mm'},
                             },
                    }
         _result_2 = checking_stack(stack=stack_2)
         self.assertTrue(_result_2)
-        
+
     def test_checking_stack_raises_value_error_if_element_missing_from_database(self):
-        '''assert checking_stack raises an error if elements can not be found in database'''
+        """assert checking_stack raises an error if elements can not be found in database"""
         stack_2 = {'Ne': {'elements': ['Ne'],
                           'stoichiometric_ratio': [1],
                           'thickness': {'value': 0.03,
                                         'units': 'mm'},
                           },
-                       'GdEu': {'elements': ['Gd','Eu'],
-                                'stoichiometric_ratio': [1,1],
-                                'thickness': {'value': 0.025,
-                                              'units': 'mm'},
-                                },
-                       }
-        self.assertRaises(ValueError, checking_stack, stack=stack_2)
-        
-    def test_checking_stack_raises_value_error_if_thickness_has_wrong_format(self):
-        '''assert checking_stack raises an error if thickness is not a number'''
-        stack_2 = {'Ag': {'elements': ['Ag'],
-                          'stoichiometric_ratio': [1],
-                          'thickness': {'value': '0.025',
-                                        'units': 'mm'},
-                          },
[…]
