### 2017-09-22 413b5b8
Changed version before deploying to pypi
