### 2017-09-23 3bd600a
Forced the points to be a integer
added to ignore some local .idea
