### 2017-09-26 335d7d5
working on auto notebook docs
