+++ docs/source/conf.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 #
-# ImagingReso documentation build configuration file, created by
+# ImagingReso docs build configuration file, created by
 # sphinx-quickstart on Wed Jul 26 14:03:30 2017.
 #
 # This file is execfile()d with the current directory set to its
@@ -15,7 +15,7 @@
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
-# documentation root, use os.path.abspath to make it absolute, like shown here.
+# docs root, use os.path.abspath to make it absolute, like shown here.
 #
 # import os
 # import sys
@@ -24,14 +24,19 @@ import sphinx_rtd_theme
 
 # -- General configuration ------------------------------------------------
 
-# If your documentation needs a minimal Sphinx version, state it here.
+# If your docs needs a minimal Sphinx version, state it here.
 #
 # needs_sphinx = '1.0'
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.viewcode']
+extensions = ['sphinx.ext.autodoc',
+              # 'sphinx.ext.viewcode',
+              'sphinx.ext.githubpages',
+              'nbsphinx',
+              'sphinx.ext.mathjax',
+              'IPython.sphinxext.ipython_console_highlighting']
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
@@ -40,14 +45,14 @@ templates_path = ['_templates']
 # You can specify multiple suffix as a list of string:
 #
 # source_suffix = ['.rst', '.md']
-source_suffix = '.rst'
+source_suffix = ['.rst', '.ipynb']
 
 # The master toctree document.
 master_doc = 'index'
 
 # General information about the project.
 project = 'ImagingReso'
-copyright = '2017, Yuxuan (Shawn) Zhang, Jean Bilheux'
+copyright = '2017, Oak Ridge National Laboratory'
 author = 'Yuxuan (Shawn) Zhang, Jean Bilheux'
 
 # The version info for the project you're documenting, acts as replacement for
@@ -59,7 +64,7 @@ version = '1.0'
 # The full version, including alpha/beta/rc tags.
 release = '0'
 
-# The language for content autogenerated by Sphinx. Refer to documentation
+# The language for content autogenerated by Sphinx. Refer to docs
 # for a list of supported languages.
 #
 # This is also used if you do content translation via gettext catalogs.
@@ -69,7 +74,7 @@ language = None
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This patterns also effect to html_static_path and html_extra_path
-exclude_patterns = []
+exclude_patterns = ['_build', '**.ipynb_checkpoints']
 
 # The name of the Pygments (syntax highlighting) style to use.
 pygments_style = 'sphinx'
@@ -80,14 +85,14 @@ todo_include_todos = False
 
 # -- Options for HTML output ----------------------------------------------
 
-# The theme to use for HTML and HTML Help pages.  See the documentation for
+# The theme to use for HTML and HTML Help pages.  See the docs for
 # a list of builtin themes.
 #
 html_theme = 'sphinx_rtd_theme'
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
-# documentation.
+# docs.
 #
 # html_theme_options = {}
 
