+++ documentation/build/html/_sources/tutorial.rst.txt
@@ -1,409 +1,0 @@
-********
-Tutorial
-********
-
-In a first time, you need to install the library
-
-$ pip install ImagingReso
-
-then you need to import it
-
->>> import ImagingReso
-
-Initialization
-##############
-
-we first define our stack of elements. Each layer of the stack can be a single element, or a compound and
-is defined by a dictionary  where the
-thickness is defined in mm, the density in g/cm3 and the ratio is the stochiometric coefficient of each element. 
-
-example:
---------
-
->>> _stack = {'CoAg': {'elements': ['Co','Ag'],
-                       'stoichiometric_ratio': [1,1],
-                       'thickness': {'value': 0.025,
-                                     'units': 'mm'},
-                       'density': {'value': 0.5,
-                                   'units': 'g/cm3'},
-                       },
-              'U': {'elements': ['U'],
-                    'stoichiometric_ratio': [1],
-                    'thickness': {'value': 0.3,
-                                  'units': 'mm'},
-                    'density': {'value': np.NaN,
-                                'units': 'g/cm3'},
-                   },
-             }
-         
-Then you can now initialize the object as followed, in this case we use a energy range of 0 to 300 eV with 
-10 eV of energy step.
-
-As noted that if the density is omitted, the program will use the stochiometri_ratio and theoretical density of each element to 
-estimage the density of the compound (layer), and changing the isotope density, or isotopic coefficient will affect the layer/compound density. 
-But if you provide the density, this value won't be affected by any other changes. If you provide the density, we suppose that you 
-know what you are doing and that you know your component enough to make sure its density value should not be changed.
-
->>> o_reso = ImagingReso.Resonance(stack = _stack, energy_min=0.001, energy_max=300, energy_step=10)
-
-It is also possible to define the layers (stack) one by one using their formula as demonstrated here
-
->>> o_reso = ImagingReso.Resonance(energy_min=0.001, energy_max=300, energy_step=10)
->>> stack1 = 'CoAg'
->>> thickness1 = 0.025 #mm
->>> density1 = 0.5 #g/cm3
->>> o_reso.add_layer(formula=stack1, thickness=thickness1, density=density1)
->>> stack2 = 'U'
->>> thcikness2 = 0.3
->>> o_reso.add_layer(formula=stack2, thcikness=thickness2)
-
-**All the parameters defined can be checked as followed**
-
-The list of stack displays the input information, but also reported the list of isotopes, mass, etc, for
-the elements you defined, for each layer.
-
->>> import pprint
->>> pprint.pprint(o_reso.slack)
-{'CoAg': {'Ag': {'density': {'units': 'g/cm3', 'value': 10.5},
-                 'isotopes': {'density': {'units': 'g/cm3',
-                                          'value': [10.406250187729098,
-                                                    10.600899412431097]},
-                              'file_names': ['Ag-107.csv', 'Ag-109.csv'],
-                              'isotopic_ratio': [0.51839, 0.48161000000000004],
-                              'list': ['107-Ag', '109-Ag'],
-                              'mass': {'units': 'g/mol',
-                                       'value': [106.905093, 108.904756]}},
-                 'molar_mass': {'units': 'g/mol', 'value': 107.8682}},
-          'Co': {'density': {'units': 'g/cm3', 'value': 8.9},
-                 'isotopes': {'density': {'units': 'g/cm3',
[…]
