+++ documentation/build/html/_sources/index.rst.txt
@@ -1,35 +1,0 @@
-.. ImagingReso documentation master file, created by
-   sphinx-quickstart on Wed Jul 26 14:03:30 2017.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-.. image:: https://img.shields.io/pypi/v/ImagingReso.svg
-  :target: https://pypi.python.org/pypi/ImagingReso
-
-.. image:: https://travis-ci.org/ornlneutronimaging/ImagingReso.svg?branch=master
-  :target: https://travis-ci.org/ornlneutronimaging/ImagingReso
-    
-.. image:: https://codecov.io/gh/ornlneutronimaging/ImagingReso/branch/master/graph/badge.svg
-  :target: https://codecov.io/gh/ornlneutronimaging/ImagingReso
-  
-.. image:: https://readthedocs.org/projects/imagingreso/badge/?version=latest
-  :target: http://imagingreso.readthedocs.io/en/latest/?badge=latest
-  :alt: Documentation Status
-  
-    
-Welcome to ImagingReso's documentation!
-=======================================
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-   
-   introduction
-   tutorial
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
