+++ documentation/source/introduction.rst
@@ -1,5 +1,0 @@
-************
-Introduction
-************
-
-Describe here what the library is doing!
