+++ docs/source/index.rst
@@ -26,6 +26,8 @@ Welcome to ImagingReso's documentation!
    
    introduction
    tutorial
+   acknowledgement
+   references
 
 Indices and tables
 ==================
