+++ docs/source/acknowledgement.rst
@@ -1,0 +1,9 @@
+****************
+Acknowledgements
+****************
+
+This work is sponsored by the Laboratory Directed Research and
+Development Program of Oak Ridge National Laboratory, managed by
+UT-Battelle LLC, for DOE. Part of this research is supported by the U.S.
+Department of Energy, Office of Science, Office of Basic Energy
+Sciences, User Facilities under contract number DE-AC05-00OR22725.
