+++ docs/source/readthedocs-environment.yml
@@ -1,0 +1,15 @@
+name: ImagingReso
+channels:
+  - conda-forge
+dependencies:
+  - python==3.5
+  - pandoc
+  - nbformat
+  - jupyter_client
+  - ipython
+  - nbconvert
+  - sphinx>=1.5.1
+  - ipykernel
+  - sphinx_rtd_theme
+  - pip:
+    - nbsphinx
