### 2017-09-27 b5123e0
renamed docs to documentation
updated supported python version in README.rst and readthedocs.io
