+++ documentation/source/conf.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 #
-# ImagingReso docs build configuration file, created by
+# ImagingReso documentation build configuration file, created by
 # sphinx-quickstart on Wed Jul 26 14:03:30 2017.
 #
 # This file is execfile()d with the current directory set to its
@@ -15,7 +15,7 @@
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
-# docs root, use os.path.abspath to make it absolute, like shown here.
+# documentation root, use os.path.abspath to make it absolute, like shown here.
 #
 # import os
 # import sys
@@ -24,7 +24,7 @@ import sphinx_rtd_theme
 
 # -- General configuration ------------------------------------------------
 
-# If your docs needs a minimal Sphinx version, state it here.
+# If your documentation needs a minimal Sphinx version, state it here.
 #
 # needs_sphinx = '1.0'
 
@@ -64,7 +64,7 @@ version = '1.0'
 # The full version, including alpha/beta/rc tags.
 release = '0'
 
-# The language for content autogenerated by Sphinx. Refer to docs
+# The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
 #
 # This is also used if you do content translation via gettext catalogs.
@@ -85,14 +85,14 @@ todo_include_todos = False
 
 # -- Options for HTML output ----------------------------------------------
 
-# The theme to use for HTML and HTML Help pages.  See the docs for
+# The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
 html_theme = 'sphinx_rtd_theme'
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
-# docs.
+# documentation.
 #
 # html_theme_options = {}
 
