+++ documentation/source/index.rst
@@ -27,7 +27,6 @@ Welcome to ImagingReso's documentation!
    introduction
    tutorial
    acknowledgement
-   references
 
 Indices and tables
 ==================
