+++ docs/source/references.rst
@@ -1,36 +1,0 @@
-**********
-References
-**********
-
-[1] M. B. Chadwick et al., “ENDF/B-VII.1 Nuclear Data for Science and
-Technology: Cross Sections, Covariances, Fission Product Yields and
-Decay Data,” Nuclear Data Sheets, vol. 112, no. 12, pp. 2887–2996, Dec.
-2011.
-
-[2] T. E. Oliphant, “SciPy: Open Source Scientific Tools for Python,”
-Computing in Science and Engineering, vol. 9. pp. 10–20, 2007.
-
-[3] S. van der Walt et al., “The NumPy Array: A Structure for Efficient
-Numerical Computation,” Computing in Science & Engineering, vol. 13, no.
-2, pp. 22–30, Mar. 2011.
-
-[4] J. D. Hunter, “Matplotlib: A 2D Graphics Environment,” Computing in
-Science & Engineering, vol. 9, no. 3, pp. 90–95, May 2007.
-
-[5] W. McKinney, “Data Structures for Statistical Computing in Python,”
-in Proceedings of the 9th Python in Science Conference, 2010, pp. 51–56.
-
-[6] P. A. Kienzle, “Periodictable V1.5.0,” Journal of Open Source
-Software, Jan. 2017.
-
-[7] M. Ooi et al., “Neutron Resonance Imaging of a Au-In-Cd Alloy for
-the JSNS,” Physics Procedia, vol. 43, pp. 337–342, 2013.
-
-[8] A. S. Tremsin et al., “Non-Contact Measurement of Partial Gas
-Pressure and Distribution of Elemental Composition Using Energy-Resolved
-Neutron Imaging,” AIP Advances, vol. 7, no. 1, p. 15315, 2017.
-
-[9] Y. Zhang et al., “The Nature of Electrochemical Delithiation of
-Li-Mg Alloy Electrodes: Neutron Computed Tomography and Analytical
-Modeling of Li Diffusion and Delithiation Phenomenon,” Journal of the
-Electrochemical Society, vol. 164, no. 2, pp. A28–A38, 2017.
