### 2017-10-03 8fc325c
adding export function as requested
