+++ ImagingReso/resonance.py
@@ -573,8 +573,15 @@ class Resonance(object):
                 [['CoAg','Ag','107-Ag'], ['CoAg']]
             if the dictionary is empty, everything is plotted
         """
+        if x_axis not in ['energy', 'lambda', 'time']:
+            raise ValueError("Please specify the x-axis type using one from '['energy', 'lambda', 'time']'.")
+        if time_unit not in ['s', 'us', 'ns']:
+            raise ValueError("Please specify the time unit using one from '['s', 'us', 'ns']'.")
+
+        # figure size
         plt.figure(figsize=[8, 8])
 
+        # stack from self
         _stack_signal = self.stack_signal
         _stack = self.stack
         _x_axis = self.total_signal['energy_eV']
@@ -604,12 +611,17 @@ class Resonance(object):
                 _x_axis = 1e9 * _utilities.ev_to_s(array=_x_axis,
                                                    source_to_detector_m=source_to_detector_m,
                                                    offset_us=offset_us)
+            print("'{}' was obtained with the following:\nsource_to_detector_m={}\noffset_us={}"
+                  .format(x_axis_label, source_to_detector_m, offset_us))
+
         if x_axis == 'number':
             x_axis_label = 'Image number (#)'
             _x_axis = _utilities.ev_to_image_number(array=_x_axis,
                                                     source_to_detector_m=source_to_detector_m,
                                                     offset_us=offset_us,
                                                     time_resolution_us=time_resolution_us)
+            print("'{}' was obtained with the following:\nsource_to_detector_m={}\noffset_us={}\ntime_resolution_us={}"
+                  .format(x_axis_label, source_to_detector_m, offset_us, time_resolution_us))
 
         """Y-axis"""
         # determine to plot transmission or attenuation
@@ -659,3 +671,135 @@ class Resonance(object):
         plt.ylabel(y_axis_label)
         plt.legend(loc='best')
         plt.show()
+
+    # def export(self, filename='simulation.csv', to_csv=False,
+    #            transmission=False, x_axis='energy',
+    #            mixed=True,
+    #            all_layers=False,
+    #            all_elements=False,
+    #            all_isotopes=False,
+    #            items_to_plot=[], time_unit='us',
+    #            offset_us=2.99, time_resolution_us=0.16, source_to_detector_m=16.125):
+    #     """
+    #     Output x and y values to .csv file
+    #     display the transmission or attenuation of compound, element and/or isotopes specified
+    #
+    #     Parameters:
+    #     ===========
+    #     transmission: boolean. True -> display transmission signal
+    #                            False -> display the attenuation signal
+    #     x_axis: string. Must be either 'energy' or 'lambda'
+    #     mixed: boolean. True -> display the total of each layer
+    #                     False -> not displayed
+    #     all_layers: boolean. True -> display all layers
+    #                          False -> not displayed
+    #     all_elements: boolean. True -> display all elements signal
+    #                            False -> not displayed
+    #     all_isotopes: boolean. True -> display all isotopes signal
+    #                            False -> not displayed
+    #     items_to_plot: array that descibes what to plot
+    #         ex:
+    #             [['CoAg','Ag','107-Ag'], ['CoAg']]
+    #         if the dictionary is empty, everything is plotted
+    #     filename:
+    #     angstrom: bool to determine the output x
+    #     transmission: bool to determine the output y
+    #
+    #     return: .csv file
+    #     """
[…]
