### 2017-10-04 4c61df6
export function added
