+++ ImagingReso/adding_export.py
@@ -28,10 +28,10 @@ o_reso.add_layer(formula=_layer_1, thickness=_thickness_1, density=_density_1)
 o_reso.add_layer(formula=_layer_2, thickness=_thickness_2)
 o_reso.add_layer(formula=_layer_3, thickness=_thickness_3, density=_density_3)
 
-o_reso.plot(mixed=True, transmission=True)
-o_reso.plot(x_axis='lambda', mixed=True, transmission=True)
-o_reso.plot(mixed=True, all_layers=True, transmission=False)
-o_reso.plot(mixed=True, all_elements=True, transmission=False)
-o_reso.plot(mixed=True, all_isotopes=True, transmission=False)
-o_reso.plot(items_to_plot=[['UO3'], ['UO3', 'U', '235-U'], ['UO3', 'U', '238-U']])
-o_reso.plot(items_to_plot=[['Co'], ['Ag'], ['Ag', 'Ag', '107-Ag'], ['Co', 'Co', '58-Co']])
+# o_reso.export(to_csv=True, mixed=True, transmission=True)
+# o_reso.export(x_axis='lambda', mixed=True, transmission=True)
+# o_reso.export(mixed=True, all_layers=True, transmission=False)
+# o_reso.export(mixed=True, all_elements=True, transmission=False)
+# o_reso.export(mixed=True, all_isotopes=True, transmission=False)
+o_reso.export(items_to_export=[['UO3'], ['UO3', 'U', '233-U'], ['UO3', 'U', '238-U']])
+# o_reso.export(items_to_plot=[['Co'], ['Ag'], ['Ag', 'Ag', '107-Ag'], ['Co', 'Co', '58-Co']])
