+++ ImagingReso/resonance.py
@@ -3,6 +3,7 @@ import numbers
 import os
 import matplotlib.pyplot as plt
 import json
+import pandas as pd
 
 from ImagingReso import _utilities
 
@@ -559,7 +560,7 @@ class Resonance(object):
         ===========
         transmission: boolean. True -> display transmission signal
                                False -> display the attenuation signal
-        x_axis: string. Must be either 'energy' or 'lambda'
+        x_axis: string. Must be either 'energy' or 'lambda' or 'time' or 'number'
         mixed: boolean. True -> display the total of each layer
                         False -> not displayed
         all_layers: boolean. True -> display all layers
@@ -568,13 +569,13 @@ class Resonance(object):
                                False -> not displayed
         all_isotopes: boolean. True -> display all isotopes signal
                                False -> not displayed
-        items_to_plot: array that descibes what to plot
+        items_to_plot: array that describes what to plot
             ex:
                 [['CoAg','Ag','107-Ag'], ['CoAg']]
             if the dictionary is empty, everything is plotted
         """
-        if x_axis not in ['energy', 'lambda', 'time']:
-            raise ValueError("Please specify the x-axis type using one from '['energy', 'lambda', 'time']'.")
+        if x_axis not in ['energy', 'lambda', 'time', 'number']:
+            raise ValueError("Please specify the x-axis type using one from '['energy', 'lambda', 'time', 'number']'.")
         if time_unit not in ['s', 'us', 'ns']:
             raise ValueError("Please specify the time unit using one from '['s', 'us', 'ns']'.")
 
@@ -585,6 +586,7 @@ class Resonance(object):
         _stack_signal = self.stack_signal
         _stack = self.stack
         _x_axis = self.total_signal['energy_eV']
+        x_axis_label = None
 
         """X-axis"""
         # determine values and labels for x-axis with options from
@@ -594,7 +596,7 @@ class Resonance(object):
         if x_axis == 'lambda':
             x_axis_label = u"Wavelength (\u212B)"
             _x_axis = _utilities.ev_to_angstroms(array=_x_axis)
-            # plt.xlim(0, 1)
+            plt.xlim(0, 1)
         if x_axis == 'time':
             if time_unit == 's':
                 x_axis_label = 'Time (s)'
@@ -622,6 +624,8 @@ class Resonance(object):
                                                     time_resolution_us=time_resolution_us)
             print("'{}' was obtained with the following:\nsource_to_detector_m={}\noffset_us={}\ntime_resolution_us={}"
                   .format(x_axis_label, source_to_detector_m, offset_us, time_resolution_us))
+        if x_axis_label is None:
+            raise ValueError("x_axis_label does NOT exist, please check.")
 
         """Y-axis"""
         # determine to plot transmission or attenuation
@@ -672,134 +676,139 @@ class Resonance(object):
         plt.legend(loc='best')
         plt.show()
 
-    # def export(self, filename='simulation.csv', to_csv=False,
-    #            transmission=False, x_axis='energy',
-    #            mixed=True,
-    #            all_layers=False,
-    #            all_elements=False,
-    #            all_isotopes=False,
-    #            items_to_plot=[], time_unit='us',
-    #            offset_us=2.99, time_resolution_us=0.16, source_to_detector_m=16.125):
-    #     """
-    #     Output x and y values to .csv file
-    #     display the transmission or attenuation of compound, element and/or isotopes specified
-    #
-    #     Parameters:
-    #     ===========
-    #     transmission: boolean. True -> display transmission signal
-    #                            False -> display the attenuation signal
-    #     x_axis: string. Must be either 'energy' or 'lambda'
-    #     mixed: boolean. True -> display the total of each layer
[…]
