### 2017-10-04 b2c83f4
added a flag to skip items_to_plot or items_to_export
