### 2017-10-04 #407 @arokem
@nicoguaro : have you had a chance to take a look at the other items on the check list above?
