### 2017-10-05 724ac82
added the support to export elemental or isotopic sigmas for making plots
