### 2017-10-05 c87cf01
updated tutorial of exporting sigma
