### 2017-10-05 #407 @nicoguaro
Yes, I installed the software for Python 3 as described in the instructions.

I also try to run the tests using ``pytest`` and this is the errors I get

        =========================================== ERRORS ===========================================
        ____________________ ERROR collecting tests/ImagingReso/utilities_test.py ____________________
        ImportError while importing test module '/home/nicoguaro/workspace/ImagingReso/tests/ImagingReso/utilities_test.py'.
        Traceback:
        utilities_test.py:7: in <module>
            from ImagingReso.resonance import Resonance
        E   ImportError: No module named ImagingReso.resonance
        !!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 1 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!

and

    =========================================== ERRORS ===========================================
    ____________________ ERROR collecting tests/ImagingReso/resonance_test.py ____________________
    ImportError while importing test module '/home/nicoguaro/workspace/ImagingReso/tests/ImagingReso/resonance_test.py'.
    Traceback:
    resonance_test.py:4: in <module>
        from ImagingReso.resonance import Resonance
    E   ImportError: No module named ImagingReso.resonance
    !!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 1 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!
    ================================== 1 error in 0.23 seconds ===================================
