### 2017-10-05 3f2c45a
added options to select max lambda for display and updated docs
