### 2017-10-05 #407 @zhangy6x
@nicoguaro I was able to reproduce the error. Look like it is the path issues with the pytest. Could you please try to run`python -m pytest tests/` in the repo dir. It fixes this problem on my end.
