### 2017-10-06 c8ce49e
added unit test and removed repeated ones. updated tutorial.
