### 2017-10-06 dec90a2
passed unit test, removed commented parts.
