+++ ImagingReso/_utilities.py
@@ -545,74 +545,3 @@ def ev_to_image_number(offset_us, source_to_detector_m, time_resolution_us, t_st
     time_record_us = (time_tot_us + offset_us)
     image_number = (time_record_us - t_start_us) / time_resolution_us
     return image_number
-
-
-# def convert_x_axis(array=[], from_units='ev', to_units='Angstroms',
-#                    offset_us=np.NaN,
-#                    source_to_detector_m=np.NaN):
-#     """allow to convert the x-axis into eV, Angstroms units, or s
-#     Parameters:
-#     ===========
-#     array: array to convert
-#     from_units: string (default is eV)
-#     to_units: string (default is Angstroms)
-#     offset_us: float. Delay in micros of the detector
-#     source_to_detector_m: float. distance source detector in m
-#     Returns:
-#     ========
-#     converted array
-#     """
-#     units_allowed = ['ev', 'angstroms', 's']
-#
-#     if array == []:
-#         return np.ndarray([])
-#
-#     _array = array.copy()
-#
-#     from_units = from_units.lower()
-#     if not from_units in units_allowed:
-#         raise ValueError("Units {} not supported!".format(from_units))
-#
-#     to_units = to_units.lower()
-#     if not to_units in units_allowed:
-#         raise ValueError("Units {} not supported!".format(to_units))
-#
-#     if from_units == to_units:
-#         return array
-#
-#     if from_units == 'ev':
-#         if to_units == 'angstroms':
-#             return ev_to_angstroms(array=array)
-#
-#         if to_units == 's':
-#             if np.isnan(source_to_detector_m):
-#                 raise ValueError("Please provide a distance source-detector in m!")
-#
-#             return ev_to_s(array=array,
-#                            offset_us=offset_us,
-#                            source_to_detector_m=source_to_detector_m)
-#
-#     if from_units == 'angstroms':
-#         if to_units == 'ev':
-#             return angstroms_to_ev(array=array)
-#
-#         if to_units == 's':
-#             if np.isnan(source_to_detector_m):
-#                 raise ValueError("Please provide a distance source-detector in m!")
-#
-#             return angstroms_to_s(array=array,
-#                                   offset_us=offset_us,
-#                                   source_to_detector_m=source_to_detector_m)
-#
-#     if from_units == 's':
-#         if to_units == 'ev':
-#             return s_to_ev(array=array,
-#                            offset_us=offset_us,
-#                            source_to_detector_m=source_to_detector_m)
-#
-#         if to_units == 'angstroms':
-#             return s_to_angstroms(array=array,
-#                                   offset_us=offset_us,
-#                                   source_to_detector_m=source_to_detector_m)
-#
-#     return np.ndarray([])
