+++ setup.py
@@ -12,7 +12,7 @@ def read_file(filename):
 
 setup(
     name = "ImagingReso",
-    version = "1.5.10",
+    version = "1.5.11",
     author = "Yuxuan (Shawn) Zhang, Jean Bilheux",
     author_email = "zhangy6@ornl.gov, bilheuxjm@ornl.gov",
     packages = find_packages(exclude=['tests', 'notebooks']),
