### 2017-10-16 #407 @zhangy6x
@nicoguaro Were you able to run the unit test with `python -m pytest tests/`?
