### 2017-10-17 #407 @nicoguaro
I was not. I obtain the following error

    _____________ ERROR collecting tests/ImagingReso/resonance_test.py _____________
    ../../anaconda2/lib/python2.7/site-packages/_pytest/python.py:418: in _importtestmodule
        mod = self.fspath.pyimport(ensuresyspath=importmode)
    ../../anaconda2/lib/python2.7/site-packages/py/_path/local.py:662: in pyimport
        __import__(modname)
    ../../anaconda2/lib/python2.7/site-packages/_pytest/assertion/rewrite.py:216: in load_module
        py.builtin.exec_(co, mod.__dict__)
    tests/ImagingReso/resonance_test.py:4: in <module>
        from ImagingReso.resonance import Resonance
    E     File "ImagingReso/resonance.py", line 115
    E       self.stack = {**self.stack, **new_stack}
    E                      ^
    E   SyntaxError: invalid syntax
