### 2017-10-17 #407 @zhangy6x
@nicoguaro I was able to reproduce your error. Could you please try to run it again under Python 3.6 instead of 2.7? If you would like, you could also find the test results on Travis (https://travis-ci.org/ornlneutronimaging/ImagingReso). The `tutorial.ipynb` in notebook dir can also be used to test all functionalities provided. Thanks.
