### 2017-10-17 18e98f7
make travis to test on linux
