### 2017-10-18 #407 @zhangy6x
@nicoguaro I wasn't able to reproduce this error. After reading the log, `['Co-59.csv', 'Co-58.csv'] != ['Co-58.csv', 'Co-59.csv']` and `['111-Ag', '110-Ag', '107-Ag', '109-Ag'] != ['107-Ag', '109-Ag', '110-Ag', '111-Ag']`, it looks like that the tests failed due to the descending sequence of the elements arranged in the lists returned. I noticed that you are using a Linux machine, so I added Linux to my Travis setting and it passed all the tests as shown below. I am not sure why you got the error you after runing the tests. Maybe the other author, @JeanBilheux, could help us to investigate this.
```============================= test session starts ==============================
platform linux -- Python 3.6.2, pytest-3.2.3, py-1.4.34, pluggy-0.4.0 -- /home/travis/mc/envs/testenv/bin/python
cachedir: .cache
rootdir: /home/travis/build/ornlneutronimaging/ImagingReso, inifile:
plugins: cov-2.5.1
collected 61 items
tests/ImagingReso/resonance_test.py::TestInitialization::test_E_max PASSED
tests/ImagingReso/resonance_test.py::TestInitialization::test_E_min PASSED
tests/ImagingReso/resonance_test.py::TestInitialization::test_adding_layer PASSED
tests/ImagingReso/resonance_test.py::TestInitialization::test_correct_initialization_of_stack PASSED
tests/ImagingReso/resonance_test.py::TestInitialization::test_element_metadata_via_add_layer_initialization PASSED
tests/ImagingReso/resonance_test.py::TestInitialization::test_element_metadata_via_stack_initialization PASSED
tests/ImagingReso/resonance_test.py::TestInitialization::test_get_sigma_isotopes PASSED
tests/ImagingReso/resonance_test.py::TestInitialization::test_initialization_E_step_bigger_than_E_range_raises_error PASSED
tests/ImagingReso/resonance_test.py::TestInitialization::test_initialization_of_object PASSED
tests/ImagingReso/resonance_test.py::TestInitialization::test_initialization_with_same_E_min_and_max_raises_error PASSED
tests/ImagingReso/resonance_test.py::TestInitialization::test_layer_density_locked_if_defined_during_initialization_add_layer PASSED
tests/ImagingReso/resonance_test.py::TestInitialization::test_layer_density_locked_if_defined_during_initialization_init PASSED
tests/ImagingReso/resonance_test.py::TestInitialization::test_repr PASSED
tests/ImagingReso/resonance_test.py::TestInitialization::test_str PASSED
tests/ImagingReso/resonance_test.py::TestGetterSetter::test_density_returned PASSED
tests/ImagingReso/resonance_test.py::TestGetterSetter::test_density_returned_all PASSED
tests/ImagingReso/resonance_test.py::TestGetterSetter::test_if_element_misssing_use_compound_and_raises_error_if_wrong PASSED
tests/ImagingReso/resonance_test.py::TestGetterSetter::test_if_element_misssing_use_compound_and_raises_error_if_wrong_when_getting_density PASSED
tests/ImagingReso/resonance_test.py::TestGetterSetter::test_list_all_stoichiometric_ratio PASSED
tests/ImagingReso/resonance_test.py::TestGetterSetter::test_retrieve_density_raises_error_if_unknown_compound PASSED
tests/ImagingReso/resonance_test.py::TestGetterSetter::test_retrieve_stoichiometric_of_UO3_sample PASSED
tests/ImagingReso/resonance_test.py::TestGetterSetter::test_retrieve_stoichiometric_ratio_raises_error_if_unknown_compound PASSED
tests/ImagingReso/resonance_test.py::TestGetterSetter::test_set_stoichiometric_ratio_correctly_calculates_new_density PASSED
tests/ImagingReso/resonance_test.py::TestGetterSetter::test_set_stoichiometric_ratio_correctly_calculates_new_molar_mass PASSED
tests/ImagingReso/resonance_test.py::TestGetterSetter::test_set_stoichiometric_ratio_has_correct_new_list_size PASSED
tests/ImagingReso/resonance_test.py::TestGetterSetter::test_set_stoichiometric_ratio_raises_value_error_if_wrong_compound PASSED
tests/ImagingReso/resonance_test.py::TestGetterSetter::test_set_stoichiometric_ratio_raises_value_error_if_wrong_element PASSED
tests/ImagingReso/resonance_test.py::TestGetterSetter::test_stoichiometric_ratio_returned PASSED
tests/ImagingReso/resonance_test.py::TestTransmissionAttenuation::test_calculate_attenuation_compound PASSED
tests/ImagingReso/resonance_test.py::TestTransmissionAttenuation::test_calculate_attenuation_element PASSED
tests/ImagingReso/resonance_test.py::TestTransmissionAttenuation::test_calculate_attenuation_isotopes PASSED
tests/ImagingReso/resonance_test.py::TestTransmissionAttenuation::test_calculate_transmission_compound PASSED
tests/ImagingReso/resonance_test.py::TestTransmissionAttenuation::test_calculate_transmission_element PASSED
tests/ImagingReso/resonance_test.py::TestTransmissionAttenuation::test_calculate_transmission_isotopes PASSED
tests/ImagingReso/resonance_test.py::TestPlot::test_axis_type_and_time_unit PASSED
tests/ImagingReso/resonance_test.py::TestExport::test_axis_type_and_time_unit PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_1::test_checking_stack PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_1::test_checking_stack_raises_value_error_if_element_missing_from_database PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_1::test_checking_stack_raises_value_error_if_thickness_has_wrong_format PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_1::test_formula_to_dictionary_raises_error_when_unknow_element PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_1::test_formula_to_dictionary_works_with_various_cases PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_1::test_get_compound_density PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_1::test_get_density PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_1::test_get_isotope_dicts_returns_correct_dictionary PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_1::test_get_isotope_dicts_returns_correct_element PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_1::test_get_isotope_returns_empty_dict_if_missing_element PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_1::test_get_list_of_element_raise_error_if_wrong_database PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_1::test_get_mass PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_1::test_is_element_in_database PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_1::test_raises_error_when_size_stoichiometric_ratio_is_different_from_elements PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_2::test_angstroms_s_conversion PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_2::test_calculate_transmission PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_2::test_energy_angstroms_conversion PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_2::test_ev_s_conversion PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_2::test_ev_to_image_number_conversion PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_2::test_get_atoms_per_cm3_of_layer PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_2::test_get_database_name_raises_error_if_wrong_file PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_2::test_get_database_works PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_2::test_get_interpolated_data PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_2::test_get_sigma PASSED
tests/ImagingReso/utilities_test.py::TestUtilities_2::test_set_distance_units PASSED
----------- coverage: platform linux, python 3.6.2-final-0 -----------
Name                                  Stmts   Miss  Cover
---------------------------------------------------------
tests/ImagingReso/resonance_test.py     324      0   100%
tests/ImagingReso/utilities_test.py     216      0   100%
---------------------------------------------------------
TOTAL                                   540      0   100%
========================== 61 passed in 12.24 seconds ==========================```
