### 2017-10-18 #407 @nicoguaro
@zhangy6x, I also tried running the notebook in the repo, and obtained this:

    <ipython-input-25-7b2e6a77814b> in <module>()
    ----> 1 energy_UO3_U_238_angstroms = o_reso.convert_x_axis(array=energy_UO3_U_238U, from_units='ev', to_units='angstroms')

    AttributeError: 'Resonance' object has no attribute 'convert_x_axis'
