### 2017-10-18 #407 @zhangy6x
@nicoguaro I am sorry, a while after submitting this code, function `o_reso.convert_x_axis` has been replaced by another keywords argument (`x_axis=`, 'energy' or 'lambda' or 'time' or 'number'. Type of x-axis.) in `o_reso.plot()` function. Could you please install the latest version of *ImagingReso* and use the notebook from the current repo to test? Sorry for any inconveniences.
