### 2017-10-18 28b7303
make travis to test both linux and mac
