+++ .travis.yml
@@ -1,8 +1,13 @@
 sudo: false
 # dist: trusty
-os:
-  - linux
-  - osx
+
+matrix:
+  include:
+    - os: linux
+      dist: trusty
+      sudo: required
+    - os: osx
+      osx_image: xcode7.2
 
 language: python
 
