### 2017-10-18 2878d09
changed the test setting back. trusty seems gives error
