+++ .travis.yml
@@ -1,14 +1,6 @@
 sudo: false
 # dist: trusty
 
-matrix:
-  include:
-    - os: linux
-      dist: trusty
-      sudo: required
-    - os: osx
-      osx_image: xcode7.2
-
 language: python
 
 python:
