### 2017-10-18 #407 @nicoguaro
I reinstalled ImagingReso and recloned the repo.

This is the test report:

    rootdir: /home/nguarinz/workspace/ImagingReso, inifile:
    collected 61 items

    tests/ImagingReso/resonance_test.py ..F.................................
    tests/ImagingReso/utilities_test.py .......F.................

    =================================== FAILURES ===================================
    _____________________ TestInitialization.test_adding_layer _____________________

    self = <resonance_test.TestInitialization testMethod=test_adding_layer>

        def test_adding_layer(self):
            """assert adding_layer works"""
            o_reso = Resonance()

            # layer 1
            layer1 = 'CoAg'
            thickness1 = 0.025
            o_reso.add_layer(formula=layer1, thickness=thickness1)

            # layer 2
            layer2 = 'Ag'
            thickness2 = 0.1
            density2 = 0.5
            o_reso.add_layer(formula=layer2, thickness=thickness2, density=density2)

            returned_stack = o_reso.stack

            expected_stack = {'CoAg': {'elements': ['Co', 'Ag'],
                                       'stoichiometric_ratio': [1, 1],
                                       'thickness': {'value': 0.025,
                                                     'units': 'mm'},
                                       'density': {'value': 9.7,
                                                   'units': 'g/cm3'},
                                       'Co': {'isotopes': {'file_names': ['Co-58.csv', 'Co-59.csv'],
                                                           'list': ['58-Co', '59-Co'],
                                                           'mass': {'value': [57.9357576, 58.9332002],
                                                                    'units': 'g/mol',
                                                                    },
                                                           'isotopic_ratio': [0.0, 1.0],
                                                           },
                                              'density': {'value': 8.9,
                                                          'units': 'g/cm3'},
                                              'molar_mass': {'value': 58.9332,
                                                             'units': 'g/mol'},
                                              },
                                       'Ag': {'isotopes': {'file_names': ['Ag-107.csv', 'Ag-109.csv'],
                                                           'list': ['107-Ag', '109-Ag'],
                                                           'mass': {'value': [106.905093, 108.904756],
                                                                    'units': 'g/mol',
                                                                    },
                                                           'isotopic_ratio': [0.51839, 0.48161000],
                                                           },
                                              'density': {'value': 10.5,
                                                          'units': 'g/cm3'},
                                              'molar_mass': {'value': 107.8682,
                                                             'units': 'g/cm3'},
                                              },
                                       },
                              'Ag': {'elements': ['Ag'],
                                     'stoichiometric_ratio': [1],
                                     'thickness': {'value': 0.1,
                                                   'units': 'mm'},
                                     'density': {'value': 0.5,
                                                 'units': 'g/cm3'},
                                     'Ag': {'isotopes': {'file_names': ['Ag-107.csv', 'Ag-109.csv'],
                                                         'list': ['107-Ag', '109-Ag'],
                                                         'mass': {'value': [106.905093, 108.904756],
                                                                  'units': 'g/mol',
                                                                  },
                                                         'isotopic_ratio': [0.51839, 0.48161000],
                                                         },
                                            'density': {'value': 10.5,
                                                        'units': 'g/cm3'},
                                            'molar_mass': {'value': 107.8682,
                                                           'units': 'g/mol'},
                                            },
                                     },
                              }

            # elements
            self.assertEqual(returned_stack['CoAg']['elements'], expected_stack['CoAg']['elements'])
            # atomic_atomic_ratio
            self.assertEqual(returned_stack['CoAg']['stoichiometric_ratio'], expected_stack['CoAg']['stoichiometric_ratio'])
            # thickness
            self.assertEqual(returned_stack['CoAg']['thickness']['value'], expected_stack['CoAg']['thickness']['value'])
            self.assertEqual(returned_stack['CoAg']['thickness']['units'], expected_stack['CoAg']['thickness']['units'])
            # file names
            self.assertEqual(returned_stack['CoAg']['Co']['isotopes']['file_names'],
    >                        expected_stack['CoAg']['Co']['isotopes']['file_names'])
    E       AssertionError: Lists differ: ['Co-59.csv', 'Co-58.csv'] != ['Co-58.csv', 'Co-59.csv']
    E
    E       First differing element 0:
    E       'Co-59.csv'
    E       'Co-58.csv'
    E
    E       - ['Co-59.csv', 'Co-58.csv']
    E       + ['Co-58.csv', 'Co-59.csv']

    tests/ImagingReso/resonance_test.py:243: AssertionError
    ______ TestUtilities_1.test_get_isotope_dicts_returns_correct_dictionary _______

    self = <utilities_test.TestUtilities_1 testMethod=test_get_isotope_dicts_returns_correct_dictionary>

        def test_get_isotope_dicts_returns_correct_dictionary(self):
            """assert get_isotope_dict works with typical entry element Ag"""
            _element = 'Ag'
            _dict_returned = get_isotope_dicts(element=_element)
            _dict_expected = {'density': {'value': 10.5,
                                          'units': 'g/cm3'},
                              'isotopes': {'list': ['107-Ag', '109-Ag', '110-Ag', '111-Ag'],
                                           'file_names': ['Ag-107.csv', 'Ag-109.csv', 'Ag-110.csv', 'Ag-111.csv'],
                                           'mass': {'value': [106.905093, 108.904756, 109.90611, 110.905295],
                                                    'units': 'g/mol'},
                                           'density': {'value': [10.406, 10.600, 10.698, 10.795],
                                                       'units': 'g/cm3'},
                                           'isotopic_ratio': [0.51839, 0.481610, 0.0, 0.0]},
                              'molar_mass': {'value': 107.8682,
                                             'units': 'g/cm3'},
                              }

            # list of isotopes
    >       self.assertEqual(_dict_returned['isotopes']['list'], _dict_expected['isotopes']['list'])
    E       AssertionError: Lists differ: ['111-Ag', '110-Ag', '107-Ag', '109-Ag'] != ['107-Ag', '109-Ag', '110-Ag', '111-Ag']
    E
    E       First differing element 0:
    E       '111-Ag'
    E       '107-Ag'
    E
    E       - ['111-Ag', '110-Ag', '107-Ag', '109-Ag']
    E       + ['107-Ag', '109-Ag', '110-Ag', '111-Ag']

    tests/ImagingReso/utilities_test.py:178: AssertionError
    ===================== 2 failed, 59 passed in 8.16 seconds ======================
