### 2017-10-18 #407 @zhangy6x
@nicoguaro Thanks a lot for reinstalling everything! This test error is same as the previous issue of descending arrangement of the elements in the lists returned `['Co-59.csv', 'Co-58.csv'] != ['Co-58.csv', 'Co-59.csv']` and `['111-Ag', '110-Ag', '107-Ag', '109-Ag'] != ['107-Ag', '109-Ag', '110-Ag', '111-Ag']`. This test works okay on my osx machine and linux on travis. I don't know why it failed. @JeanBilheux, could you please help me with this? Thanks.
