### 2017-10-18 50d895d
automatically sorting list of isotopes files retrieved.
