### 2017-10-18 #407 @JeanBilheux
@nicoguaro I updated the code to force the sorting of the Isotopes files. This should allow the unit test to work on any platform.
