### 2017-10-18 e3da5e8
passed unit test, removed commented parts.
