### 2017-10-18 #407 @nicoguaro
@JeanBilheux, should I reinstall then?

    pip install --upgrade ImagingReso

would make the trick?
