### 2017-10-18 #407 @nicoguaro
I needed to git pull the repository, but got the tests to pass. Below is the output

    rootdir: /home/nguarinz/workspace/ImagingReso, inifile:
    collected 61 items

    tests/ImagingReso/resonance_test.py ....................................
    tests/ImagingReso/utilities_test.py .........................

    ========================== 61 passed in 7.99 seconds ===========================
