### 2017-10-18 #407 @nicoguaro
@zhangy6x, the tests are working now in my machine. I will add details of the review partially. Below are some comments:

- The project presents a statement of need of the software. Nevertheless, I think that it can be further improved for a wider audience.
- The project includes data in the repository. As far as I understand, the database is in the public domain, but I would like the authors to confirm this point.
- There are not community guidelines in the repository.
-  I ran the IPython notebook and it worked this time. A little detail at the very end is the following when I try to export data as instructed in the previous cell.

with output:

    KeyError                                  Traceback (most recent call last)
    <ipython-input-43-b67aeaf1b993> in <module>()
          1 o_reso.export(y_axis='attenuation', items_to_export=[['CoAg', 'Co'],['CoAg','Ag','107-Ag']],
    ----> 2              filename="resonance.csv", to_csv=True)

    ~/anaconda2/envs/ImagingReso/lib/python3.6/site-packages/ImagingReso/resonance.py in export(self, filename, to_csv, x_axis, y_axis, mixed, all_layers, all_elements, all_isotopes, items_to_export, time_unit, offset_us, time_resolution_us, source_to_detector_m, t_start_us)
        769                     while _path_to_export:
        770                         _item = _path_to_export.pop(0)
    --> 771                         _live_path = _live_path[_item]
        772                     _y_axis = _live_path[y_axis_tag]
        773                     df[_label] = _y_axis

    KeyError: 'CoAg'

 Also, it is not clear to me why do you have to tell the method ``export`` explicitly that you want to write the export file. Is there a reason behind this?

@arokem, is it OK to make suggestions regarding things like PEP8?
