### 2017-10-19 #407 @zhangy6x
@nicoguaro Thanks a lot for your comments.

- We are very happy that you find this package could be further improved for a wider audience. When we started writing this package, it was mainly for simulating the resonance signal expected from neutron imaging measurement. Of course, we are happy to further improve it in later versions as you suggested.
- Yes, we did receive the positive reply from the NNDC database owner, David Brown (email: dbrown@bnl.gov, Brookhaven National Laboratory), of including the database in our package. And we did cite related ENDF publication from his group.
- I didn't notice that the community guidelines are required for this submission. I am happy to add them into `README.rst` once I discussed it with @JeanBilheux.
- I am sorry, I just updated the tutorial and now it is showing the correct instruction regarding exporting specified items. The new version will be available for `pip install` once Jean released it on PyPi tomorrow.
- The `export` function here is mainly for providing options if the user would like to construct their own plot with the signal obtained from this package. Since the resulted resonance signal is fairly complex, instead of exporting the whole thing, the user can choose different types of the x-axis (energy or wavelength of neutrons) and the y-axis (attenuation, transmission or cross-sections) to make their own graph conveniently.
