### 2017-10-19 #407 @JeanBilheux
@nicoguaro

* Can you be more specific about "improve for a wider audience"? Do you mean X_ray community? other? What kind of improvements do you think about?
* Same thing for the "community guidelines", what is this?
