### 2017-10-19 #407 @nicoguaro
@JeanBilheux

- I can try, at least. With wider audience I mean ... wider. I don't think that the software will be used by a person with zero knowledge of the topic. But if the description is good enough for any person to understand what is does, it does no harm. Of course, It is really difficult to write something that "any person" can understand, but I would bet for something broader than a specific community.
- Community guidelines refer to explicitly state how to contribute to the project. The importance of this is based on the openness of the software. Check the [author guidelines](http://joss.theoj.org/about#author_guidelines), for further detail.
