### 2017-10-19 fb3c65d
replaced to_csv in o_reso.export
