### 2017-10-20 b932112
added community guidelines
