### 2017-10-21 #407 @zhangy6x
@nicoguaro Community guidelines have been added to README. Thanks.
