### 2017-10-24 #407 @nicoguaro
@arokem, any guidelines regarding formatting?
