### 2017-10-29 #407 @arokem
@nicoguaro : do you mean the formatting of the paper? Or the community guidelines?

Also, @zhangy6x : I noticed that you have added a notice of copyright to the paper. Could you please explain what it means, and why you need it? We usually license these articles under a Creative Commons license, and this might be in conflict (?)
