### 2017-10-29 #407 @arokem
In particular, JOSS does not claim any copyright, and the work published is licensed under CC-BY.
That means that *anyone* can reproduce the work (under the CC-BY license), and you can especially do so because you retain copyright on the work.

So - would you mind removing that notice of copyright please?

Other than that - I think that the paper is formatted as needed
