### 2017-10-29 #407 @nicoguaro
@arokem, no, I was referring to the code. Is it okay to suggest to make some changes regarding formatting?
