### 2017-10-30 #407 @JeanBilheux
@arokem   Our program is using some data from the Brookhaven database (listed in the md file). Those are available publicly but to be sure we contacted one of the people in charge of the database. He told us that we were allowed to use the data as long as we were acknowledging them for the data. In the future, we are planning on moving this code into the scikit-beam project, and this is the reason we matched our license with their package.
