### 2017-10-30 #407 @arokem
@JeanBilheux : You can certainly acknowledge the use of the Brookhaven database (and I see that you are referring to it in the text of your paper)! But I was referring to the section before the acknowledgements (titled "notice of copyright"). You cannot set the copyright on the paper in this way.
