### 2017-10-30 #407 @arokem
@nicoguaro : you can certainly make comments about the formatting of the code, but the authors might not be required to comply with your comments as a requirement for publication. We do not require that code be formatted in a particular way as a condition for publication.
