### 2017-10-30 #407 @nicoguaro
@arokem, I understand that they are not required to comply. But it was better to ask before suggesting, I think.

@zhangy6x, I suggest to check the [PEP8](https://www.python.org/dev/peps/pep-0008/) guidelines for style in Python in the sourcecode.
