### 2017-10-31 #407 @JeanBilheux
@arokem Sorry, I got it now. I'm currently checking this with my management and our lawyers here at the lab (to play it safe). I will get back to you soon.
