### 2017-11-02 da0da46
fixed pep8 and added plot sigma functionality
