+++ ImagingReso/resonance.py
@@ -10,8 +10,8 @@ from ImagingReso import _utilities
 class Resonance(object):
     database = 'ENDF_VIII'
 
-    E_MIN = 1e-5
-    E_MAX = 3e3
+    e_min = 1e-5
+    e_max = 3e3
 
     stack = {}  # compound, thickness, atomic_ratio of each layer with isotopes information
     stack_sigma = {}  # all the energy and sigma of the isotopes and compounds
@@ -26,11 +26,11 @@ class Resonance(object):
 
     def __init__(self, stack={}, energy_max=1, energy_min=0.001, energy_step=0.001):
         """initialize resonance object
-        
+
         Parameters:
         ==========
-        stack: dictionary 
-          example: {'layer1': {'elements':['Ag','Si], 
+        stack: dictionary
+          example: {'layer1': {'elements':['Ag','Si],
                                'atomic_ratio': [1, 2],
                                'thickness': {'value': 0.025,
                                              'units': 'mm',
@@ -45,12 +45,12 @@ class Resonance(object):
         """
         self.__element_metadata = {}
 
-        if energy_min < self.E_MIN:
-            raise ValueError("Energy min (eV) must be >= {}".format(self.E_MIN))
+        if energy_min < self.e_min:
+            raise ValueError("Energy min (eV) must be >= {}".format(self.e_min))
         self.energy_min = energy_min
 
-        if energy_max > self.E_MAX:
-            raise ValueError("Energy max (eV) must be <= {}".format(self.E_MAX))
+        if energy_max > self.e_max:
+            raise ValueError("Energy max (eV) must be <= {}".format(self.e_max))
         self.energy_max = energy_max
 
         if energy_min == energy_max:
@@ -75,7 +75,7 @@ class Resonance(object):
 
     def __str__(self):
         """what to display if user does
-        
+
         >>> o_reso = Resolution()
         >>> print(o_reso)
         """
@@ -83,7 +83,7 @@ class Resonance(object):
 
     def __repr__(self):
         """what to display if user does
-        
+
         >>> o_reso = Resolution()
         >>> o_reso
         """
@@ -91,13 +91,13 @@ class Resonance(object):
 
     def add_layer(self, formula='', thickness=np.NaN, density=np.NaN):
         """provide another way to define the layers (stack)
-        
+
         Parameters:
         ===========
         formula: string
            ex: 'CoAg2'
            ex: 'Al'
-        thickness: float (in mm) 
+        thickness: float (in mm)
         density: float (g/cm3)
         """
         if formula == '':
@@ -118,12 +118,12 @@ class Resonance(object):
 
     def get_isotopic_ratio(self, compound='', element=''):
         """returns the list of isotopes for the element of the compound defined with their stoichiometric values
-        
+
         Parameters:
         ===========
         compound: string (default is empty). If empty, all the stoichiometric will be displayed
-        element: string (default is same as compound). 
-        
+        element: string (default is same as compound).
+
         Raises:
         =======
         ValueError if element is not defined in the stack
@@ -146,7 +146,7 @@ class Resonance(object):
 
         # checking compound is valid
         list_compounds = _stack.keys()
-        if not compound in list_compounds:
+        if compound not in list_compounds:
             list_compounds_joined = ', '.join(list_compounds)
             raise ValueError("Compound '{}' could not be find in {}".format(compound, list_compounds_joined))
 
@@ -155,7 +155,7 @@ class Resonance(object):
             # we assume that the element and compounds names matched
             element = compound
         list_element = _stack[compound].keys()
-        if not element in list_element:
+        if element not in list_element:
             list_element_joined = ', '.join(list_element)
             raise ValueError("Element '{}' should be any of those elements: {}".format(element, list_element_joined))
 
[…]
