### 2017-11-03 caec66d
added sigma_raw to stack_sigma
