### 2017-11-03 #407 @JeanBilheux
@arokem Here is the reply I got from the management/lawyers.

"""
ORNL’s prime contract with DOE requires the copyright in acknowledgement statements. The notice is a Prime Contract requirement.  Specifically, Paragraph I.141(d)(2) of our Prime Contract states: “The contractor shall mark each scientific or technical article first produced or composed under this Contract and submitted for journal publication or similar means of dissemination with a notice, similar in all material respects to the following, on the front reflecting the Government's non-exclusive, paid-up, irrevocable, world-wide license in the copyright.”

I verified with ORNL’s IP counsel that the CC-BY license is not in conflict with the government’s retained rights in the publication.

I would suggest that you respond that, “Our prime contract with DOE requires this statement.  In addition, the CC-BY license is not in conflict with the government’s retained rights in the publication.”
"""
