### 2017-11-09 #407 @arokem
After some conversation with other JOSS editors, I am starting to understand that the copyright notice certainly is not in conflict with the CC-BY license (I was wrong in stating this before), but it is also really not necessary, and your paper will be less confusing to readers if you remove it. This is because, as the contract you cite from states, you are required to use a notice that is "similar in all material respects to the following". The CC-BY license is already that - anyone can reproduce the article, and certainly the United States government (or anyone designated by US government) can do so. I believe that readers will be un-neccesarily confused by this notice (as was I).
