### 2017-11-15 #407 @zhangy6x
@arokem Thanks a lot for your help throughout this review process. Please excuse us for this belated response since I don't have much knowledge about copyright issues.
After discussion, we are glad that this copyright notice is not conflicting with the CC-BY license, and we think it is better to keep it based on the Prime Contract requirement. However, we could move this notice to a more appropriate place to avoid any confusion. If that is the case, please advise us where should this notice be kept. Thanks much!
