### 2017-11-15 #407 @zhangy6x
@nicoguaro PEP8 check and formatting has been performed as you suggested. Thanks.
