### 2017-11-15 #407 @nicoguaro
@arokem, I consider that this review is ready to go.
