### 2017-11-16 #407 @arokem
@zhangy6x : how about moving this from its own section to the "acknowledgements" section?
