### 2017-11-16 56069b6
reallocate copyright notice to acknowledgements.
