+++ ImagingReso/plot_various_x_axis.py
@@ -1,5 +1,4 @@
 from ImagingReso.resonance import Resonance
-import numpy as np
 import pprint
 # Global parameters
 _energy_min = 1
