### 2017-11-16 3e5e0d6
Delete plot_various_x_axis.py
