### 2017-11-16 e4f37ae
reallocate copyright notice to acknowledgements.
