+++ README.rst
@@ -125,21 +125,6 @@ C\ :sub:`i` : molar concentration of element *i*,
 
 m\ :sub:`ij` : atomic mass values for the isotope *j* of element *i*.
 
-Acknowledgements
-----------------
-
-This work is sponsored by the Laboratory Directed Research and
-Development Program of Oak Ridge National Laboratory, managed by
-UT-Battelle LLC, under Contract No. DE-AC05-00OR22725 with the U.S.
-Department of Energy. The United States Government retains and the
-publisher, by accepting the article for publication, acknowledges
-that the United States Government retains a non-exclusive, paid-up,
-irrevocable, worldwide license to publish or reproduce the published
-form of this manuscript, or allow others to do so, for United States
-Government purposes. The Department of Energy will provide public
-access to these results of federally sponsored research in accordance
-with the DOE Public Access Plan(http://energy.gov/downloads/doe-public-access-plan).
-
 References
 ----------
 
@@ -187,6 +172,18 @@ Distributed under the BSD license. See ``LICENSE.txt`` for more information
 
 https://github.com/ornlneutronimaging/ImagingReso
 
+Acknowledgements
+----------------
 
-
+This work is sponsored by the Laboratory Directed Research and
+Development Program of Oak Ridge National Laboratory, managed by
+UT-Battelle LLC, under Contract No. DE-AC05-00OR22725 with the U.S.
+Department of Energy. The United States Government retains and the
+publisher, by accepting the article for publication, acknowledges
+that the United States Government retains a non-exclusive, paid-up,
+irrevocable, worldwide license to publish or reproduce the published
+form of this manuscript, or allow others to do so, for United States
+Government purposes. The Department of Energy will provide public
+access to these results of federally sponsored research in accordance
+with the DOE Public Access Plan(http://energy.gov/downloads/doe-public-access-plan).
 
