### 2017-11-16 #407 @zhangy6x
@arokem : Notice of Copyright has been moved to "acknowledgments" section. Thanks!
