### 2017-11-16 bed9376
update doc to be consistent
