+++ documentation/source/acknowledgement.rst
@@ -4,6 +4,12 @@ Acknowledgements
 
 This work is sponsored by the Laboratory Directed Research and
 Development Program of Oak Ridge National Laboratory, managed by
-UT-Battelle LLC, for DOE. Part of this research is supported by the U.S.
-Department of Energy, Office of Science, Office of Basic Energy
-Sciences, User Facilities under contract number DE-AC05-00OR22725.
+UT-Battelle LLC, under Contract No. DE-AC05-00OR22725 with the U.S.
+Department of Energy. The United States Government retains and the
+publisher, by accepting the article for publication, acknowledges
+that the United States Government retains a non-exclusive, paid-up,
+irrevocable, worldwide license to publish or reproduce the published
+form of this manuscript, or allow others to do so, for United States
+Government purposes. The Department of Energy will provide public
+access to these results of federally sponsored research in accordance
+with the DOE Public Access Plan(http://energy.gov/downloads/doe-public-access-plan).
