### 2017-11-16 #407 @arokem
@arfon: I believe this paper is ready for publication
