### 2017-11-17 #407 @arfon
@zhangy6x - Could you make sure you cite the references you currently have in the file `paper.bib` file in the markdown paper (`paper.md`) (You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html))
