+++ paper/paper.md
@@ -20,17 +20,26 @@ bibliography: paper.bib
 
 # Summary
 
-ImagingReso is an open-source Python library that simulates the neutron 
-resonance signal for neutron imaging measurements. By defining the sample 
-information such as density, thickness in the neutron path, and isotopic 
-ratios of the elemental composition of the material, this package plots 
-the expected resonance peaks for a selected neutron energy range. Various 
-sample types such as layers of single elements (Ag, Co, etc. in solid form), 
-chemical compounds (UO<sub>3</sub>, Gd<sub>2</sub>O<sub>3</sub>, etc.), 
-or even multiple layers of both types can be plotted with this package. 
-Major plotting features include display of the transmission/attenuation 
-in wavelength, energy, and time scale, and show/hide elemental and isotopic 
-contributions in the total resonance signal.
+ImagingReso is an open-source Python library that simulates the neutron
+resonance signal for neutron imaging measurements. By defining the sample
+information such as density, thickness in the neutron path, and isotopic
+ratios of the elemental composition of the material, this package plots
+the expected resonance peaks for a selected neutron energy range.
+Various sample types such as layers of single elements (Ag, Co, etc. in solid form),
+chemical compounds (UO<sub>3</sub>, Gd<sub>2</sub>O<sub>3</sub>, etc.),
+or even multiple layers of both types can be plotted with this package.
+Major plotting features include display of the transmission/attenuation in
+wavelength, energy, and time scale, and show/hide elemental and
+isotopic contributions in the total resonance signal.
+
+The energy dependent cross-section data used in this library are from
+[National Nuclear Data Center](http://www.nndc.bnl.gov/), a published
+online database. [Evaluated Nuclear Data File 
+(ENDF/B)](http://www.nndc.bnl.gov/exfor/endf00.jsp) [@Chadwick2011] is currently
+supported and more evaluated databases will be added in future.
+
+Python packages used are: SciPy [@Oliphant2007], NumPy [@Stéfan van der Walt2011], 
+Matplotlib [@Hunter2007], Pandas [@McKinney2010] and Periodictable [@kienzle_p_a_2017_840347].
 
 The energy dependent cross-section data used in this library are from 
 [National Nuclear Data Center](http://www.nndc.bnl.gov/), an online database 
@@ -38,11 +47,32 @@ published by Brookhaven National Laboratory.
 [Evaluated Nuclear Data File (ENDF/B)](http://www.nndc.bnl.gov/exfor/endf00.jsp) 
 is currently supported and more evaluated databases will be added in the future.
 
-# References
-paper.bib
+The neutron transmission calculation algorithm of neutron transmission *T*(*E*), 
+is base on Beer-Lambert law [@Ooi2013;@Tremsin2017;@Zhang2017]:
+
+<img src="documentation/source/_static/Beer_lambert_law_1.png" width=504 height=80 />
+
+*N<sub>i</sub>* : number of atoms per unit volume of element *i*, 
+
+*d<sub>i</sub>* : effective thickness along the neutron path of element *i*,
+
+*σ<sub>ij</sub>(E)* : energy-dependent neutron total cross-section for the isotope *j* of element *i*, 
+
+*A<sub>ij</sub>* : abundance for the isotope *j* of element *i*. 
+
+For solid materials the number of atoms per unit volume can be calculated from:
+
+<img src="documentation/source/_static/Beer_lambert_law_2.png" width=244 height=80 />
+
+*N<sub>A</sub>* : Avogadro’s number,
+
+*C<sub>i</sub>* : molar concentration of element *i*,
+
+*ρ<sub>i</sub>* : density of the element *i*,
+
[…]
