+++ paper/paper.md
@@ -51,9 +51,7 @@ is currently supported and more evaluated databases will be added in the future.
 The neutron transmission calculation algorithm of neutron transmission *T*(*E*), 
 is base on Beer-Lambert law [@Ooi2013;@Tremsin2017;@Zhang2017]:
 
-![Figure_1](https://github.com/ornlneutronimaging/ImagingReso/blob/master/documentation/source/_static/Beer_lambert_law_1.png)
-
-![Figure_1](http://latex.codecogs.com/gif.latex?%24%24T%5Cleft%28%20E%20%5Cright%29%20%3D%5Cfrac%20%7B%20I%5Cleft%28%20E%20%5Cright%29%20%7D%7B%20%7B%20I%20%7D_%7B%200%20%7D%5Cleft%28%20E%20%5Cright%29%20%7D%20%3Dexp%5Cleft%5B%20-%5Csum%5Cnolimits_i%20%7B%20%7B%20N%20%7D_%7B%20i%20%7D%7B%20d%20%7D_%7B%20i%20%7D%20%7D%20%5Csum%5Cnolimits_j%20%7B%20%7B%20%5Csigma%20%7D_%7B%20ij%20%7D%5Cleft%28%20E%20%5Cright%29%20%7B%20A%20%7D_%7B%20ij%20%7D%20%7D%20%5Cright%5D%24%24)
+<img src="https://github.com/ornlneutronimaging/ImagingReso/blob/master/documentation/source/_static/Beer_lambert_law_1.png" width=504 height=80 />
 
 *N<sub>i</sub>* : number of atoms per unit volume of element *i*, 
 
@@ -65,7 +63,7 @@ is base on Beer-Lambert law [@Ooi2013;@Tremsin2017;@Zhang2017]:
 
 For solid materials the number of atoms per unit volume can be calculated from:
 
-![Figure_2](https://github.com/ornlneutronimaging/ImagingReso/blob/master/documentation/source/_static/Beer_lambert_law_2.png)
+<img src="https://github.com/ornlneutronimaging/ImagingReso/blob/master/documentation/source/_static/Beer_lambert_law_2.png" width=244 height=80 />
 
 *N<sub>A</sub>* : Avogadro’s number,
 
