+++ paper/paper.bib
@@ -13,7 +13,7 @@ volume = {9},
 year = {2007}
 }
 
-@article{Stéfan van der Walt2011,
+@article{Stefan2011,
 author = {Stéfan van der Walt and S. Chris Colbert and Gaël Varoquaux},
 title = {The NumPy Array: A Structure for Efficient Numerical Computation},
 journal = {Computing in Science \& Engineering},
