### 2017-11-17 #407 @zhangy6x
@arfon : I have updated `paper.md` and created a new archive on Zenodo (DOI: 10.5281/zenodo.1054038). Thanks much.
