+++ paper/paper.bib
@@ -1,5 +1,4 @@
 @misc{Oliphant2007,
-abstract = {By itself, Python is an excellent "steering" language for scientific codes written in other languages. However, with additional basic tools, Python transforms into a high-level language suited for scientific and engineering code that's often fast enough to be immediately useful but also flexible enough to be sped up with additional extensions.},
 annote = {[Online; accessed ]},
 author = {Oliphant, Travis E},
 booktitle = {Computing in Science and Engineering},
@@ -35,19 +34,18 @@ number = {3},
 pages = {90-95},
 year = {2007},
 doi = {10.1109/MCSE.2007.55},
-abstract =  {Matplotlib is a 2D graphics package used for Python for application development, interactive scripting, and publication-quality image generation across user interfaces and operating systems.},
-publisher = {IEEE COMPUTER SOC}
+publisher = {IEEE COMPUTER SOC},
 URL = {http://aip.scitation.org/doi/abs/10.1109/MCSE.2007.55},
 eprint = {http://aip.scitation.org/doi/pdf/10.1109/MCSE.2007.55}
 }
 
-@InProceedings{ McKinney2010,
+@InProceedings{McKinney2010,
   author    = { Wes McKinney },
   title     = { Data Structures for Statistical Computing in Python },
   booktitle = { Proceedings of the 9th Python in Science Conference },
   pages     = { 51 - 56 },
   year      = { 2010 },
-  editor    = { St\'efan van der Walt and Jarrod Millman }
+  editor    = { Stéfan van der Walt and Jarrod Millman }
 }
 
 @misc{kienzle_p_a_2017_840347,
@@ -76,7 +74,6 @@ author = "M.B. Chadwick and M. Herman and P. Oblo{\v z}insk{\' y} and M.E. Dunn 
 }
 
 @article{Ooi2013,
-abstract = {The Japan Spallation Neutron Source (JSNS) at the Japan Proton Accelerator Research Complex (J-PARC) was developed as a 1- MW spallation neutron source. A Ag-In-Cd alloy was used as the decoupler material in two decoupled moderators. Although the Ag-In-Cd decoupler brings about superior neutronic performance, it has the disadvantage of high residual radioactivity. A Au-In- Cd alloy has been proposed as a solution to this problem. Recently, we successfully produced a ternary Au-In-Cd alloy. The alloy composition was 74.9 at{\%} Au, 0.5 at{\%} In, and 24.6 at{\%} Cd. The distribution of the elements in the alloy was first determined by energy-dispersive X-ray (EDX) analysis. However, it was difficult to measure the In distribution by EDX because the amount of In is very small, and its spectrum is similar to that of Cd. Therefore, pulsed neutron imaging using both a time gated camera system and a multi-channel plate detector was performed to measure the elements in the Au-In-Cd alloy. The analysis was performed at the BL10 in the JSNS on samples of the Au-In-Cd alloy, an In foil, and two Au foils. With this technique, the distribution of Au, In, and Cd in the Au-In-Cd specimen was distinctly determined.},
 author = {Ooi, M and Teshigawara, M and Kai, T and Harada, M and Maekawa, F and Futakawa, M and Hashimoto, E and Segawa, M and Kureta, M and Tremsin, A and Kamiyama, T and Kiyanagi, Y},
 doi = {http://dx.doi.org/10.1016/j.phpro.2013.03.040},
 isbn = {1875-3892},
@@ -103,7 +100,6 @@ eprint = {http://dx.doi.org/10.1063/1.4975632}
 }
 
 @article{Zhang2017,
[…]
