+++ paper/paper.bib
@@ -53,7 +53,6 @@ eprint = {http://aip.scitation.org/doi/pdf/10.1109/MCSE.2007.55}
 @misc{kienzle_p_a_2017_840347,
   author       = {Kienzle, P. A.},
   title        = {periodictable v1.5.0},
-  journal      = "Journal of Open Source Software",
   month        = may,
   year         = 2017,
   note         = {Funded in part by NSF DMR-0520547 and DMR-1508249},
